refactor: ClubValidator 삭제에 따른 수정
