+++ src/main/java/com/mobble/mobbleserver/refactor/like/clubLike/service/ClubLikeService.java
@@ -1,7 +1,9 @@
 package com.mobble.mobbleserver.refactor.like.clubLike.service;
 
-import com.mobble.mobbleserver.refactor.club.core.entity.Club;
-import com.mobble.mobbleserver.refactor.club.core.validator.ClubValidator;
+import com.mobble.mobbleserver.application.club.core.port.required.ClubReadPort;
+import com.mobble.mobbleserver.domain.club.core.Club;
+import com.mobble.mobbleserver.global.exception.common.DomainException;
+import com.mobble.mobbleserver.global.exception.errorCode.club.ClubErrorCode;
 import com.mobble.mobbleserver.refactor.like.baseLike.entity.LikeType;
 import com.mobble.mobbleserver.refactor.like.baseLike.service.AbstractLikeService;
 import com.mobble.mobbleserver.refactor.like.clubLike.entity.ClubLike;
@@ -17,7 +19,8 @@ import java.util.Optional;
 public class ClubLikeService extends AbstractLikeService<Club, ClubLike> {
 
     private final ClubLikeRepository clubLikeRepository;
-    private final ClubValidator clubValidator;
+
+    private final ClubReadPort clubReadPort;
 
     @Override
     public LikeType getType() {
@@ -26,7 +29,7 @@ public class ClubLikeService extends AbstractLikeService<Club, ClubLike> {
 
     @Override
     protected Club getTarget(Long targetId) {
-        return clubValidator.findClubByClubIdOrThrow(targetId);
+        return findClubByClubIdOrThrow(targetId);
     }
 
     @Override
@@ -47,5 +50,10 @@ public class ClubLikeService extends AbstractLikeService<Club, ClubLike> {
     @Override
     protected void deleteLike(ClubLike entity) {
         clubLikeRepository.delete(entity);
+    }
+
+    private Club findClubByClubIdOrThrow(Long clubId) {
+        return clubReadPort.findById(clubId)
+                .orElseThrow(() -> new DomainException((ClubErrorCode.NOT_FOUND)));
     }
 }
